Move providers to top level
